requirements.txt
allure-pytest==2.13.4
fastapi==0.110.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.4
pytest==8.0.2
pytest-clarity==1.0.1
mypy==1.9.0
curlify==2.2.1
sqlalchemy==2.0.29
pytest-xdist==3.5.0
httpx==0.27.0
uvicorn[standard]==0.29.0
curlify==2.2.1